services/vision/requirements.txt
fastapi
uvicorn
# ML Core
torch
transformers
accelerate
# Image Processing
pillow
timm
einops
# Ollama Client
ollama
# Transcribe Video's Audio
openai-whisper